server: move helper message types to backend types
